refactor: TxUserService 구현
