+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,62 +1,9 @@
 package com.techcourse.service;
 
-import com.interface21.dao.DataAccessException;
-import com.interface21.jdbc.datasource.DataSourceUtils;
-import com.interface21.transaction.support.TransactionSynchronizationManager;
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
 
-import javax.sql.DataSource;
-import java.sql.Connection;
-import java.sql.SQLException;
+public interface UserService {
 
-public class UserService {
-
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
-    private final DataSource dataSource;
-
-    public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao, final DataSource dataSource) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-        this.dataSource = dataSource;
-    }
-
-    public User findById(final long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        Connection connection = DataSourceUtils.getConnection(dataSource);
-        try {
-            connection.setAutoCommit(false);
-
-            try {
-                final var user = userDao.findById(id);
-                user.changePassword(newPassword);
-                userDao.update(user);
-                userHistoryDao.log(new UserHistory(user, createBy));
-
-                connection.commit();
-            } catch (Exception e) {
-                try {
-                    connection.rollback();
-                } catch (SQLException rollbackEx) {
-                    e.addSuppressed(rollbackEx);
-                }
-                throw new DataAccessException("Failed to change password for user id: " + id, e);
-            }
-        } catch (SQLException e) {
-            throw new DataAccessException("Failed to manage transaction", e);
-        } finally {
-            TransactionSynchronizationManager.unbindResource(dataSource);
-            DataSourceUtils.releaseConnection(connection, dataSource);
-        }
-    }
+    User findById(final long id);
+    void changePassword(final long id, final String newPassword, final String createdBy);
 }
